+++ webApplication/src/main/java/data/Dot.java
@@ -2,14 +2,14 @@ package data;
 
 public class Dot{
 
-    private final int x;
+    private final double x;
     private final double y;
-    private final int r;
+    private final double r;
 
     private boolean status;
 
 
-    public Dot(int x, double y, int r) {
+    public Dot(double x, double y, double r) {
         this.x = x;
         this.y = y;
         this.r = r;
@@ -19,7 +19,7 @@ public class Dot{
 
 
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -27,7 +27,7 @@ public class Dot{
         return y;
     }
 
-    public int getR() {
+    public double getR() {
         return r;
     }
